cleanup (TabLayout): move to `~/components/Base`
